fixed issues from review and added docs
