+++ .gitignore
@@ -3,4 +3,3 @@ release
 lib
 coverage
 .vscode
-yarn.lock
