+++ docs/API.md
@@ -728,16 +728,69 @@ const Counter = componentFromStream(props$ => {
 })
 ```
 
-### `mapPropsStreamWithConfig()`
+### `componentFromStreamWithConfig()`
+
 ```js
-mapPropsStreamWithConfig<Stream>(
+componentFromStreamWithConfig<Stream>(
   config: {
     fromESObservable<T>: ?(observable: Observable<T>) => Stream<T>,
     toESObservable<T>: ?(stream: Stream<T>) => Observable<T>,
-  },
-  ownerPropsToChildProps: (props$: Stream<object>) => Stream<object>,
-  BaseComponent: ReactElementType
+  }
+) => (
+  propsToReactNode: (props$: Stream<object>) => Stream<ReactNode>
 ): ReactComponent
+```
+
+Alternative to `componentFromStream()` that accepts an observable config and returns a customized `componentFromStream()` which uses the specified observable library. This option is recommended if you want to avoid global state with `setObservableConfig()`.
+**Note: The following configuration modules are not included in the main export. You must import them individually, as shown in the examples.**
+
+#### RxJS
+
+```js
+import rxjsConfig from 'recompose/rxjsObservableConfig'
+const componentFromStream = componentFromStreamWithConfig(rxjsConfig)
+```
+
+#### RxJS 4 (legacy)
+
+```js
+import rxjs4Config from 'recompose/rxjs4ObservableConfig'
+const componentFromStream = componentFromStreamWithConfig(rxjs4Config)
+```
+
+#### most
+
+```js
+import mostConfig from 'recompose/mostObservableConfig'
+const componentFromStream = componentFromStreamWithConfig(mostConfig)
+```
+
+#### xstream
+
+```js
+import xstreamConfig from 'recompose/xstreamObservableConfig'
+const componentFromStream = componentFromStreamWithConfig(xstreamCondfig)
+```
+
+#### Bacon
+
+```js
+import baconConfig from 'recompose/baconObservableConfig'
+const componentFromStream = componentFromStreamWithConfig(baconConfig)
+```
+
+#### Kefir
+
+```js
+import kefirConfig from 'recompose/kefirObservableConfig'
+const componentFromStream = componentFromStreamWithConfig(kefirConfig)
+```
+
+#### Flyd
+
+```js
+import flydConfig from 'recompose/flydObservableConfig'
+const componentFromStream = componentFromStreamWithConfig(flydConfig)
 ```
 
 ### `mapPropsStream()`
@@ -752,6 +805,21 @@ mapPropsStream(
 A higher-order component version of `componentFromStream()` — accepts a function that maps an observable stream of owner props to a stream of child props, rather than directly to a stream of React nodes. The child props are then passed to a base component.
 
 You may want to use this version to interoperate with other Recompose higher-order component helpers.
+
+### `mapPropsStreamWithConfig()`
+```js
+mapPropsStreamWithConfig<Stream>(
+  config: {
+    fromESObservable<T>: ?(observable: Observable<T>) => Stream<T>,
+    toESObservable<T>: ?(stream: Stream<T>) => Observable<T>,
+  },
+) => (
+  ownerPropsToChildProps: (props$: Stream<object>) => Stream<object>,
+  BaseComponent: ReactElementType
+): ReactComponent
+```
+
+Alternative to `mapPropsStream()` that accepts a observable config and returns a customized `mapPropsStream()` that uses the specified observable library. See `componentFromStreamWithConfig()` above.
 
 ```js
 const enhance = mapPropsStream(props$ => {
@@ -808,6 +876,8 @@ setObservableConfig({
   fromESObservable: Rx.Observable.from
 })
 ```
+
+**Note: `setObservableConfig()` uses global state, and could break apps if used inside a package inteded to be shared. See `componentFromStreamWithConfig()` and `mapPropsStreamWithConfig()` for alternatives for package authors.**
 
 In addition to `fromESObservable`, the config object also accepts `toESObservable`, which converts a stream back into an ES observable. Because RxJS 5 observables already conform to the ES observable spec, `toESObservable` is not necessary in the above example. However, it is required for libraries like RxJS 4 or xstream, whose streams do not conform to the ES observable spec.
