+++ src/packages/recompose/mostObservableConfig.js
@@ -2,6 +2,7 @@ import { from, Stream } from 'most'
 
 const config = {
   fromESObservable: from || Stream.from,
+  toESObservable: stream => stream,
 }
 
 export default config
